requirements.txt
PyPDF2>=3.0.0
requests>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
openai>=0.27.0
streamlit>=1.24.1